backend/requirements.txt
fastapi
uvicorn
web3
eth-account
python-dotenv
sqlalchemy
psycopg2-binary
requests